Profile Page now appears properly #24
Made some slight changes to the context dictionary and the templates so
that when a user has no school or hub, the correct message is shown.
Also made some tweaks to the style to make the pages more consistant
with each other.  Main changes are putting everything in hero-units (now
with less padding), and using h2 for main titles.