Adding sessionstorage saving to resource/packs
When creating/evolving/editing resources, and wen creating/editing packs, the filed in fields are stored in sessionstorage whilst the user goes to the tag page to enter more tags. when thye return to the original page, the fields are automatically filled in.  As before, sessionstorage is cleared when the form is submitted.